# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlanDay {
  id: number;
  week: number;
  phase: string;
  date: string;
  day: string;
  strengthLoad?: number | null;
  equipment: string;
  description: string;
  cardioMin?: number | null;
  distanceMi?: number | null;
  pace?: string | null;
  sessionType: string;
  isRest: boolean;
  totalLoad: number;
}

export interface PlanWeek {
  week: number;
  phase: string;
  startDate: string;
  endDate: string;
  plannedStrength?: number | null;
  plannedCardio?: number | null;
  plannedTotalLoad: number;
  plannedMiles: number;
  longRunMi: number;
  actualMiles?: number | null;
  completedSessions?: number | null;
  totalSessions?: number | null;
}

export type PlanWeekDetail = PlanWeek & {
  days: PlanDay[];
};

export interface PlanOverview {
  currentWeek: number;
  currentPhase: string;
  totalWeeks: number;
  weeksRemaining: number;
  raceDate: string;
  startDate: string;
  startWeight: number;
  currentWeight: number | null;
  goalWeight: number;
  weeklyMilesTarget?: number;
  longRunTarget?: number;
}

export interface Workout {
  id: number;
  planDayId?: number | null;
  date: string;
  equipment: string;
  sessionType: string;
  durationMin?: number | null;
  distanceMi?: number | null;
  pace?: string | null;
  avgHr?: number | null;
  rpe?: number | null;
  strengthLoad?: number | null;
  totalLoad?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface TodayPlan {
  date: string;
  hasPlan: boolean;
  plan?: PlanDay | null;
  loggedWorkout?: Workout | null;
}

export interface CreateWorkoutBody {
  planDayId?: number | null;
  date: string;
  equipment: string;
  sessionType: string;
  durationMin?: number | null;
  distanceMi?: number | null;
  pace?: string | null;
  avgHr?: number | null;
  rpe?: number | null;
  strengthLoad?: number | null;
  totalLoad?: number | null;
  notes?: string | null;
}

export interface UpdateWorkoutBody {
  date?: string;
  equipment?: string;
  sessionType?: string;
  durationMin?: number | null;
  distanceMi?: number | null;
  pace?: string | null;
  avgHr?: number | null;
  rpe?: number | null;
  strengthLoad?: number | null;
  totalLoad?: number | null;
  notes?: string | null;
}

export interface Measurement {
  id: number;
  date: string;
  weight?: number | null;
  lArm?: number | null;
  rArm?: number | null;
  lLeg?: number | null;
  rLeg?: number | null;
  belly?: number | null;
  chest?: number | null;
  notes?: string | null;
  createdAt: string;
}

export interface CreateMeasurementBody {
  date: string;
  weight?: number | null;
  lArm?: number | null;
  rArm?: number | null;
  lLeg?: number | null;
  rLeg?: number | null;
  belly?: number | null;
  chest?: number | null;
  notes?: string | null;
}

export interface UpdateMeasurementBody {
  date?: string;
  weight?: number | null;
  lArm?: number | null;
  rArm?: number | null;
  lLeg?: number | null;
  rLeg?: number | null;
  belly?: number | null;
  chest?: number | null;
  notes?: string | null;
}

export interface DashboardSummary {
  currentWeek: number;
  currentPhase: string;
  weekProgressPct: number;
  weeklyMilesActual: number;
  weeklyMilesPlanned: number;
  weeklyLoadActual: number;
  weeklyLoadPlanned: number;
  weeklySessionsCompleted: number;
  weeklySessionsPlanned: number;
  totalMilesAllTime: number;
  longestRunMi: number;
  weightStart: number;
  weightCurrent: number | null;
  weightGoal: number;
  weightLost: number;
  weightToGoal: number;
  adherencePct: number;
  daysToRace: number;
}

export interface WeightPoint {
  date: string;
  weight: number;
}

export interface WeeklyMileagePoint {
  week: number;
  startDate?: string;
  phase?: string;
  plannedMiles: number;
  actualMiles: number;
}

export interface EquipmentUsage {
  equipment: string;
  sessions: number;
  totalMinutes: number;
  totalLoad: number;
  totalDistance?: number;
}

export interface LongRunPoint {
  week: number;
  date?: string;
  plannedMi: number;
  actualMi: number;
}

export type ListWorkoutsParams = {
  limit?: number;
  from?: string;
  to?: string;
  equipment?: string;
};