# lib/api-zod/src/generated/types/createMeasurementBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMeasurementBody {
  date: string;
  weight?: number | null;
  lArm?: number | null;
  rArm?: number | null;
  lLeg?: number | null;
  rLeg?: number | null;
  belly?: number | null;
  chest?: number | null;
  notes?: string | null;
}