# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  currentWeek: number;
  currentPhase: string;
  weekProgressPct: number;
  weeklyMilesActual: number;
  weeklyMilesPlanned: number;
  weeklyLoadActual: number;
  weeklyLoadPlanned: number;
  weeklySessionsCompleted: number;
  weeklySessionsPlanned: number;
  totalMilesAllTime: number;
  longestRunMi: number;
  weightStart: number;
  weightCurrent: number | null;
  weightGoal: number;
  weightLost: number;
  weightToGoal: number;
  adherencePct: number;
  daysToRace: number;
}