# lib/api-zod/src/generated/types/planOverview.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface PlanOverview {
  currentWeek: number;
  currentPhase: string;
  totalWeeks: number;
  weeksRemaining: number;
  raceDate: string;
  startDate: string;
  startWeight: number;
  currentWeight: number | null;
  goalWeight: number;
  weeklyMilesTarget?: number;
  longRunTarget?: number;
}