# lib/api-zod/src/generated/types/planWeekDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanDay } from "./planDay";
import type { PlanWeek } from "./planWeek";

export type PlanWeekDetail = PlanWeek & {
  days: PlanDay[];
};