# lib/api-zod/src/generated/types/todayPlan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */
import type { PlanDay } from "./planDay";
import type { Workout } from "./workout";

export interface TodayPlan {
  date: string;
  hasPlan: boolean;
  plan?: PlanDay | null;
  loggedWorkout?: Workout | null;
}