# lib/api-zod/src/generated/types/weeklyMileagePoint.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface WeeklyMileagePoint {
  week: number;
  startDate?: string;
  phase?: string;
  plannedMiles: number;
  actualMiles: number;
}