# lib/api-zod/src/generated/types/workout.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marathon Command Center API
 * OpenAPI spec version: 0.1.0
 */

export interface Workout {
  id: number;
  planDayId?: number | null;
  date: string;
  equipment: string;
  sessionType: string;
  durationMin?: number | null;
  distanceMi?: number | null;
  pace?: string | null;
  avgHr?: number | null;
  rpe?: number | null;
  strengthLoad?: number | null;
  totalLoad?: number | null;
  notes?: string | null;
  createdAt: Date;
}